CLEANUP: Detach protocol utilization codes to protocol_util file
